Send email to a users who pays for applications. #APP-3 Fixed

+++ AppMarketPlace.web/src/main/java/co/edu/uniandes/csw/appmarketplace/services/TransactionService.java
@@ -14,8 +14,13 @@ import co.edu.uniandes.csw.appmarketplace.dtos.CartItemDTO;
 import co.edu.uniandes.csw.appmarketplace.dtos.ClientDTO;
 import co.edu.uniandes.csw.appmarketplace.dtos.TransactionDTO;
 import co.edu.uniandes.csw.appmarketplace.providers.StatusCreated;
+import co.edu.uniandes.csw.appmarketplace.utils.Emailer;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
+import java.util.Map;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 import javax.inject.Inject;
 import javax.servlet.http.HttpServletResponse;
 import javax.ws.rs.Consumes;
@@ -58,21 +63,34 @@ public class TransactionService {
     @QueryParam("maxRecords")
     private Integer maxRecords;
     private final ClientDTO client = (ClientDTO) SecurityUtils.getSubject().getSession().getAttribute("Client");
+    private Object currentUser;
 
     @POST
     @StatusCreated
     public void createPayment(TransactionDTO dto) {
+         Subject currentUser = SecurityUtils.getSubject();
+        ClientDTO client = (ClientDTO) currentUser.getSession().getAttribute("Client");
+        if(client==null){          
+            Logger.getLogger(QuestionService.class.getName()).log(Level.SEVERE, null, new Exception("User is not a registered client"));
+            return;
+        } 
+        Map<String, String> userAttributes = (Map<String, String>) currentUser.getPrincipals().oneByType(java.util.Map.class);
+
         dto.setPayer(client);
         dto.setPaymentCard(paymentCardLogic.getPaymentCards(dto.getId()));
         dto.setId(null);
         dto.setStatus("Aprobado");
+        int total =0;
+        int number=0;
         for (CartItemDTO cartItem : clientLogic.getClient(client.getId()).getCartItems()) {
             dto.setRecipient(cartItem.getApp());
             dto.setTotal((int) ((appLogic.getApp(cartItem.getApp().getId()).getPrice() - appLogic.getApp(cartItem.getApp().getId()).getDiscount()) * Long.parseLong(cartItem.getQuantity().toString())));
             TransactionLogic.createTransaction(dto);
             cartItemLogic.deleteCartItemByClient(client.getId(), cartItem.getId());
+            number++;
+            total+=(int) ((appLogic.getApp(cartItem.getApp().getId()).getPrice() - appLogic.getApp(cartItem.getApp().getId()).getDiscount()) * Long.parseLong(cartItem.getQuantity().toString()));
         }
-
+        Emailer.sendPaymentEmail(client.getName(), userAttributes.get("email"),  Integer.toString(total), new Date(),Integer.toString(number) );
     }
 
     @GET
